bot_music/requirements.txt
aiohttp==3.4.4
async-timeout==3.0.1
attrs==18.2.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
pycparser==2.19
PyNaCl==1.2.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
websockets==6.0
yarl==1.3.0
youtube_dl==2019.5.20
discord.py==1.1.1